Use NonEmpty Package
